Extract tracker adapter
